backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.6.1
anthropic
jiter==0.9.0
python-dotenv==1.0.1
pytest==7.4.4
starlette==0.36.3
langchain
langchain-anthropic
langchain-community
langchain-core
langgraph
redis==5.0.1
tiktoken==0.6.0
PyPDF2==3.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
boto3==1.34.45
aiofiles==23.2.1
aiosqlite==0.19.0
psutil
greenlet
backoff
colorama